refactor; moved changes to its own module
